requirements.txt
# Core
django==4.2.6
python-dotenv==1.0.0

# Deployment
whitenoise==6.5.0
gunicorn==21.2.0
django-debug-toolbar==4.2.0

# DB Layer
# psycopg2-binary==3.1.12
# mysqlclient==2.2.0

celery==5.3.4
redis==5.0.1
hiredis==2.2.3

django-celery-results==2.5.1
djangorestframework==3.14.0
django-api-generator==1.0.17

Pillow==10.0.1
psycopg==3.1.12

django-unfold==0.17.0